Change the v6 update process
