+++ src/UpdateScripts/v6_0/UpdateClassReferences.php
@@ -2,7 +2,7 @@
 
 namespace DoubleThreeDigital\SimpleCommerce\UpdateScripts\v6_0;
 
-use Illuminate\Support\Facades\Artisan;
+use DoubleThreeDigital\SimpleCommerce\Facades\Order;
 use Statamic\UpdateScripts\UpdateScript;
 
 class UpdateClassReferences extends UpdateScript
@@ -14,6 +14,24 @@ class UpdateClassReferences extends UpdateScript
 
     public function update()
     {
-        Artisan::call('sc:update-class-references');
+        Order::query()
+            ->where('gateway', '!=', null)
+            ->chunk(50, function ($orders) {
+                $orders->each(function ($order) {
+                    // When the gateway reference is still a class, change it to the handle.
+                    if ($order->gateway() && class_exists($order->gateway()['use'])) {
+                        $order->gateway(array_merge($order->gateway(), [
+                            'use' => $order->gateway()['use']::handle(),
+                        ]));
+
+                        $order->save();
+                    }
+
+                    // When the shipping method reference is still a class, change it to the handle.
+                    if ($order->has('shipping_method') && class_exists($order->get('shipping_method'))) {
+                        $order->set('shipping_method', $order->get('shipping_method')::handle())->saveQuietly();
+                    }
+                });
+            });
     }
 }
